refactor(cli): split plugin-skills.ts into flags/listing/reload modules (#366)
